Implementa algunas funciones

+++ obligatorio2-compgraf/SceneObject.cpp
@@ -16,9 +16,9 @@ float SceneObject::getTransmissionCoefficient()
     return transmissionCoefficient;
 }
 
-bool SceneObject::solveQuadratic(const float& a, const float& b, const float& c, float& x0, float& x1)
+bool SceneObject::solveQuadratic(const double& a, const double& b, const double& c, double& x0, double& x1)
 {
-    float discr = b * b - 4 * a * c;
+    double discr = b * b - 4.0 * a * c;
     if (discr < 0) return false;
     else if (discr == 0) x0 = x1 = -0.5 * b / a;
     else {
